+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * Shopper Insights (BETA)
   * For analytics, send `experiment` as a parameter to `getRecommendedPaymentMethods` method
-  * For analytics, send `experiment` and `buttonRank` analytic metrics to FPTI via the button presented event methods
+  * For analytics, send `experiment` and `paymentMethodsDisplayed` analytic metrics to FPTI via the button presented event methods
 
 ## 5.1.0 (2024-10-15)