+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Shopper Insights (BETA)
   * For analytics, send `experiment` as a parameter to `getRecommendedPaymentMethods` method
   * For analytics, send `experiment` and `paymentMethodsDisplayed` analytic metrics to FPTI via the button presented event methods
+* Venmo
+  * Add `VenmoClient` constructor with `appLinkReturnUri` to use App Links when redirecting back from the Venmo flow
 
 ## 5.1.0 (2024-10-15)