+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## unreleased
 
 * GooglePay
-  * Fixing a crash on API 33 devices. It is recommended that merchants not use 5.1.0 for GooglePay.
+  * Fix a crash being caused on API 33 devices. It is recommended that merchants not use 5.1.0 for GooglePay.
 * Shopper Insights (BETA)
   * For analytics, send `experiment` as a parameter to `getRecommendedPaymentMethods` method
   * For analytics, send `experiment` and `paymentMethodsDisplayed` analytic metrics to FPTI via the button presented event methods