+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+* GooglePay
+  * Fix a crash being caused on API 33 devices. It is recommended that merchants not use 5.1.0 for GooglePay.
 * Shopper Insights (BETA)
   * For analytics, send `experiment` as a parameter to `getRecommendedPaymentMethods` method
   * For analytics, send `experiment` and `paymentMethodsDisplayed` analytic metrics to FPTI via the button presented event methods